drying up the wasm client
