+++ clients/client-core/Cargo.toml
@@ -17,6 +17,7 @@ sled = { version = "0.34", optional = true }
 tap = "1.0.1"
 thiserror = "1.0.34"
 url = { version ="2.2", features = ["serde"] }
+tokio = { version = "1.21.2", features = ["time", "macros"]}
 
 # internal
 config = { path = "../../common/config" }
@@ -30,8 +31,11 @@ nymsphinx = { path = "../../common/nymsphinx" }
 pemstore = { path = "../../common/pemstore" }
 topology = { path = "../../common/topology" }
 validator-client = { path = "../../common/client-libs/validator-client", default-features = false }
+task = { path = "../../common/task" }
 
-tokio = { version = "1.21.2", features = ["time", "macros"]}
+[target."cfg(not(target_arch = \"wasm32\"))".dependencies.tokio-stream]
+version = "0.1.9"
+features = ["time"]
 
 [target."cfg(target_arch = \"wasm32\")".dependencies.wasm-bindgen-futures]
 version = "0.4"
@@ -47,8 +51,8 @@ rev = "b9d1a54ad514c2f230a026afe0dde341e98cd7b6"
 version = "0.2.4"
 features = ["futures"]
 
-[target."cfg(not(target_arch = \"wasm32\"))".dependencies.task]
-path = "../../common/task"
+#[target."cfg(not(target_arch = \"wasm32\"))".dependencies.task]
+#path = "../../common/task"
 
 [dev-dependencies]
 tempfile = "3.1.0"
