+++ clients/client-core/src/client/real_messages_control/acknowledgement_control/acknowledgement_listener.rs
@@ -70,10 +70,7 @@ impl AcknowledgementListener {
         }
     }
 
-    #[cfg(not(target_arch = "wasm32"))]
     pub(super) async fn run_with_shutdown(&mut self, mut shutdown: task::ShutdownListener) {
-        use std::time::Duration;
-
         debug!("Started AcknowledgementListener with graceful shutdown support");
 
         while !shutdown.is_shutdown() {
@@ -90,13 +87,13 @@ impl AcknowledgementListener {
                 }
             }
         }
-        tokio::time::timeout(Duration::from_secs(5), shutdown.recv())
+        #[cfg(not(target_arch = "wasm32"))]
+        tokio::time::timeout(std::time::Duration::from_secs(5), shutdown.recv())
             .await
             .expect("Task stopped without shutdown called");
         log::debug!("AcknowledgementListener: Exiting");
     }
 
-    #[cfg(target_arch = "wasm32")]
     pub(super) async fn run(&mut self) {
         debug!("Started AcknowledgementListener without graceful shutdown support");
