+++ clients/client-core/src/client/real_messages_control/acknowledgement_control/input_message_listener.rs
@@ -194,10 +194,7 @@ where
         }
     }
 
-    #[cfg(not(target_arch = "wasm32"))]
     pub(super) async fn run_with_shutdown(&mut self, mut shutdown: task::ShutdownListener) {
-        use std::time::Duration;
-
         debug!("Started InputMessageListener with graceful shutdown support");
 
         while !shutdown.is_shutdown() {
@@ -216,13 +213,13 @@ where
                 }
             }
         }
-        tokio::time::timeout(Duration::from_secs(5), shutdown.recv())
+        #[cfg(not(target_arch = "wasm32"))]
+        tokio::time::timeout(std::time::Duration::from_secs(5), shutdown.recv())
             .await
             .expect("Task stopped without shutdown called");
         log::debug!("InputMessageListener: Exiting");
     }
 
-    #[cfg(target_arch = "wasm32")]
     pub(super) async fn run(&mut self) {
         debug!("Started InputMessageListener without graceful shutdown support");
         while let Some(input_msg) = self.input_receiver.recv().await {
