+++ clients/client-core/src/client/received_buffer.rs
@@ -320,10 +320,7 @@ impl RequestReceiver {
         }
     }
 
-    #[cfg(not(target_arch = "wasm32"))]
     async fn run_with_shutdown(&mut self, mut shutdown: task::ShutdownListener) {
-        use std::time::Duration;
-
         debug!("Started RequestReceiver with graceful shutdown support");
         while !shutdown.is_shutdown() {
             tokio::select! {
@@ -342,13 +339,13 @@ impl RequestReceiver {
                 },
             }
         }
-        tokio::time::timeout(Duration::from_secs(5), shutdown.recv())
+        #[cfg(not(target_arch = "wasm32"))]
+        tokio::time::timeout(std::time::Duration::from_secs(5), shutdown.recv())
             .await
             .expect("Task stopped without shutdown called");
         log::debug!("RequestReceiver: Exiting");
     }
 
-    #[cfg(target_arch = "wasm32")]
     async fn run(&mut self) {
         debug!("Started RequestReceiver without graceful shutdown support");
 
@@ -374,10 +371,7 @@ impl FragmentedMessageReceiver {
         }
     }
 
-    #[cfg(not(target_arch = "wasm32"))]
     async fn run_with_shutdown(&mut self, mut shutdown: task::ShutdownListener) {
-        use std::time::Duration;
-
         debug!("Started FragmentedMessageReceiver with graceful shutdown support");
         while !shutdown.is_shutdown() {
             tokio::select! {
@@ -395,13 +389,13 @@ impl FragmentedMessageReceiver {
                 }
             }
         }
-        tokio::time::timeout(Duration::from_secs(5), shutdown.recv())
+        #[cfg(not(target_arch = "wasm32"))]
+        tokio::time::timeout(std::time::Duration::from_secs(5), shutdown.recv())
             .await
             .expect("Task stopped without shutdown called");
         log::debug!("FragmentedMessageReceiver: Exiting");
     }
 
-    #[cfg(target_arch = "wasm32")]
     async fn run(&mut self) {
         debug!("Started FragmentedMessageReceiver without graceful shutdown support");
 
@@ -438,7 +432,6 @@ impl ReceivedMessagesBufferController {
         }
     }
 
-    #[cfg(not(target_arch = "wasm32"))]
     pub fn start_with_shutdown(self, shutdown: task::ShutdownListener) {
         let mut fragmented_message_receiver = self.fragmented_message_receiver;
         let mut request_receiver = self.request_receiver;
