+++ clients/client-core/src/client/topology_control.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::spawn_future;
+use futures::StreamExt;
 use log::*;
 use nymsphinx::addressing::clients::Recipient;
 use nymsphinx::params::DEFAULT_NUM_MIX_HOPS;
@@ -296,14 +297,22 @@ impl TopologyRefresher {
         self.topology_accessor.is_routable().await
     }
 
-    #[cfg(not(target_arch = "wasm32"))]
     pub fn start_with_shutdown(mut self, mut shutdown: task::ShutdownListener) {
         spawn_future(async move {
             debug!("Started TopologyRefresher with graceful shutdown support");
 
+            #[cfg(not(target_arch = "wasm32"))]
+            let mut interval = tokio_stream::wrappers::IntervalStream::new(tokio::time::interval(
+                self.refresh_rate,
+            ));
+
+            #[cfg(target_arch = "wasm32")]
+            let mut interval =
+                gloo_timers::future::IntervalStream::new(self.refresh_rate.as_millis() as u32);
+
             while !shutdown.is_shutdown() {
                 tokio::select! {
-                    _ = tokio::time::sleep(self.refresh_rate) => {
+                    _ = interval.next() => {
                         self.refresh().await;
                     },
                     _ = shutdown.recv() => {
@@ -311,20 +320,25 @@ impl TopologyRefresher {
                     },
                 }
             }
-            tokio::time::timeout(Duration::from_secs(5), shutdown.recv())
+            #[cfg(not(target_arch = "wasm32"))]
+            tokio::time::timeout(std::time::Duration::from_secs(5), shutdown.recv())
                 .await
                 .expect("Task stopped without shutdown called");
             log::debug!("TopologyRefresher: Exiting");
         })
     }
 
-    #[cfg(target_arch = "wasm32")]
     pub fn start(mut self) {
-        use futures::StreamExt;
-
         spawn_future(async move {
+            #[cfg(not(target_arch = "wasm32"))]
+            let mut interval = tokio_stream::wrappers::IntervalStream::new(tokio::time::interval(
+                self.refresh_rate,
+            ));
+
+            #[cfg(target_arch = "wasm32")]
             let mut interval =
                 gloo_timers::future::IntervalStream::new(self.refresh_rate.as_millis() as u32);
+
             while let Some(_) = interval.next().await {
                 self.refresh().await;
             }
