+++ clients/client-core/src/init.rs
@@ -90,7 +90,6 @@ async fn register_with_gateway(
         gateway.owner.clone(),
         our_identity.clone(),
         timeout,
-        #[cfg(not(target_arch = "wasm32"))]
         None,
     );
     gateway_client
