+++ common/client-libs/gateway-client/Cargo.toml
@@ -26,6 +26,7 @@ network-defaults = { path = "../../network-defaults" }
 nymsphinx = { path = "../../nymsphinx" }
 pemstore = { path = "../../pemstore" }
 validator-client = { path = "../validator-client", optional = true }
+task = { path = "../../task" }
 
 
 [dependencies.tungstenite]
@@ -46,9 +47,6 @@ version = "0.14"
 
 [target."cfg(not(target_arch = \"wasm32\"))".dependencies.credential-storage]
 path = "../../credential-storage"
-
-[target."cfg(not(target_arch = \"wasm32\"))".dependencies.task]
-path = "../../task"
 
 # wasm-only dependencies
 [target."cfg(target_arch = \"wasm32\")".dependencies.wasm-bindgen]
