+++ common/client-libs/gateway-client/src/packet_router.rs
@@ -4,14 +4,12 @@
 // JS: I personally don't like this name very much, but could not think of anything better.
 // I will gladly take any suggestions on how to rename this.
 
+use crate::error::GatewayClientError;
 use futures::channel::mpsc;
 use log::*;
 use nymsphinx::addressing::nodes::MAX_NODE_ADDRESS_UNPADDED_LEN;
 use nymsphinx::params::packet_sizes::PacketSize;
-#[cfg(not(target_arch = "wasm32"))]
 use task::ShutdownListener;
-
-use crate::error::GatewayClientError;
 
 pub type MixnetMessageSender = mpsc::UnboundedSender<Vec<Vec<u8>>>;
 pub type MixnetMessageReceiver = mpsc::UnboundedReceiver<Vec<Vec<u8>>>;
@@ -23,7 +21,6 @@ pub type AcknowledgementReceiver = mpsc::UnboundedReceiver<Vec<Vec<u8>>>;
 pub struct PacketRouter {
     ack_sender: AcknowledgementSender,
     mixnet_message_sender: MixnetMessageSender,
-    #[cfg(not(target_arch = "wasm32"))]
     shutdown: Option<ShutdownListener>,
 }
 
@@ -31,12 +28,11 @@ impl PacketRouter {
     pub fn new(
         ack_sender: AcknowledgementSender,
         mixnet_message_sender: MixnetMessageSender,
-        #[cfg(not(target_arch = "wasm32"))] shutdown: Option<ShutdownListener>,
+        shutdown: Option<ShutdownListener>,
     ) -> Self {
         PacketRouter {
             ack_sender,
             mixnet_message_sender,
-            #[cfg(not(target_arch = "wasm32"))]
             shutdown,
         }
     }
@@ -86,7 +82,6 @@ impl PacketRouter {
         if !received_messages.is_empty() {
             trace!("routing 'real'");
             if let Err(err) = self.mixnet_message_sender.unbounded_send(received_messages) {
-                #[cfg(not(target_arch = "wasm32"))]
                 if let Some(shutdown) = &mut self.shutdown {
                     if shutdown.is_shutdown_poll() {
                         // This should ideally not happen, but it's ok
