+++ common/socks5/proxy-helpers/src/connection_controller.rs
@@ -254,6 +254,7 @@ impl Controller {
                 },
             }
         }
+        #[cfg(not(target_arch = "wasm32"))]
         tokio::time::timeout(Duration::from_secs(5), self.shutdown.recv())
             .await
             .expect("Task stopped without shutdown called");
