+++ common/task/Cargo.toml
@@ -9,7 +9,11 @@ edition = "2021"
 futures = "0.3"
 log = "0.4"
 thiserror = "1.0.37"
-tokio = { version = "1.21.2", features = ["macros", "signal", "time", "sync"] }
+tokio = { version = "1.21.2", features = ["macros", "sync"] }
+
+[target."cfg(not(target_arch = \"wasm32\"))".dependencies.tokio]
+version = "1.21.2"
+features = ["signal", "time"]
 
 [dev-dependencies]
 tokio = { version = "1.21.2", features = ["rt-multi-thread", "net", "signal", "test-util", "macros"] }
