+++ nym-connect/src-tauri/src/tasks.rs
@@ -1,4 +1,4 @@
-use client_core::config::GatewayEndpoint;
+use client_core::config::GatewayEndpointConfig;
 use futures::channel::mpsc;
 use std::sync::Arc;
 use tap::TapFallible;
@@ -25,7 +25,7 @@ pub enum Socks5StatusMessage {
 /// The main SOCKS5 client task. It loads the configuration from file determined by the `id`.
 pub fn start_nym_socks5_client(
     id: &str,
-) -> Result<(Socks5ControlMessageSender, StatusReceiver, GatewayEndpoint)> {
+) -> Result<(Socks5ControlMessageSender, StatusReceiver, GatewayEndpointConfig)> {
     log::info!("Loading config from file: {id}");
     let config = Socks5Config::load_from_file(Some(id))
         .tap_err(|_| log::warn!("Failed to load configuration file"))?;
